# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/shelf-click:
    post:
      operationId: recordShelfClicks
      tags: [analytics]
      summary: Record shelf outbound clicks
      description: |
        Persist a batch of "Visit brand" outbound clicks emitted by the
        "On our shelf" cards on ingredient briefs, routine pages, and
        concern guides. Designed to be called from
        `navigator.sendBeacon` (or `fetch` with `keepalive`); the
        endpoint always responds 204 so the browser never blocks
        navigation waiting on a body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShelfClickBatch"
      responses:
        "204":
          description: Accepted — clicks were persisted.
        "400":
          description: Malformed payload.
...
  /corrections:
    post:
      operationId: submitCorrection
      tags: [corrections]
      summary: Submit a reader-flagged correction
      description: |
        Public endpoint backing the "Submit a correction" form on the
        `/corrections` page (and the small CTA in the site footer of
        every page). Persists the submission to the editor inbox so
        the corrections workflow can triage and apply the fix per the
        SLA in the page's published five-rule policy (typo: 24h,
        factual: 7 days).

        The endpoint is intentionally unauthenticated — any reader
        should be able to flag an error without signing up — so abuse
        defences live at the API edge: a same-origin Origin/Referer
        allowlist, a per-IP fixed-window rate limit, and a server-side
        honeypot field (`website`) that real browsers leave blank.
        Submissions whose honeypot is filled in are silently accepted
        (HTTP 204) but never persisted, so bots can't tell they were
        rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorrectionSubmissionBody"
      responses:
        "204":
          description: Submission recorded (or silently dropped as honeypot spam).
        "400":
          description: Malformed payload.
        "403":
          description: Origin/Referer not on the allowlist.
        "429":
          description: Per-IP rate limit exceeded.
  /editor/corrections:
    get:
      operationId: listEditorCorrections
      tags: [corrections]
      summary: List reader-submitted corrections for the editor inbox
      description: |
        Editor-only inbox of reader-submitted corrections produced by
        the public `POST /corrections` form. Returns rows newest-first
        (by `createdAt`) so an editor scanning the queue every morning
        sees fresh submissions at the top, with the SLA clock derived
        client-side from `createdAt`.

        With `?status=` the list is filtered to a single workflow
        state. The default is `new`, matching the typical "what came
        in since I last checked" workflow. Pass `all` to disable the
        filter.
      parameters:
        - name: status
          in: query
          required: false
          description: Filter to a single submission status, or `all`.
          schema:
            type: string
            enum: [new, triaged, applied, dismissed, all]
            default: new
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
      security:
        - editorSession: []
      responses:
        "200":
          description: Reader-submitted corrections matching the filter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorCorrectionsList"
        "401":
          description: Missing or invalid editor session.
  /editor/corrections/{id}:
    patch:
      operationId: updateEditorCorrection
      tags: [corrections]
      summary: Update a reader-submitted correction's status or internal note
      description: |
        Updates the workflow status and/or the internal note on a
        reader-submitted correction. Both body fields are optional —
        sending just `status` advances the row through the workflow
        without touching the note, and sending just `internalNote`
        records a triage comment without changing the state. At least
        one field must be present or the request is rejected 400.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCorrectionBody"
      security:
        - editorSession: []
      responses:
        "200":
          description: Updated correction row.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrectionRecord"
        "400":
          description: Malformed body (no fields, or invalid status).
        "401":
          description: Missing or invalid editor session.
        "404":
          description: Correction not found.
...
components:
  securitySchemes:
    editorSession:
      type: apiKey
      in: cookie
      name: evidently_editor
      description: |
        HttpOnly signed session cookie. Issued by `POST /editor/sign-in`
        in exchange for the shared `EDITOR_TOKEN`, and valid for 30
        days. Endpoints without this scheme listed are public.
  schemas:
...
    CorrectionSubmissionBody:
      type: object
      properties:
        pageUrl:
          type: string
          minLength: 1
          maxLength: 2000
          description: |
            URL of the page the reader spotted the error on. Stored as
            free text so a submission about a third-party page or a
            link copied from elsewhere still records cleanly.
        description:
          type: string
          minLength: 5
          maxLength: 4000
          description: |
            Reader's description of what they think is wrong. The 5-char
            floor stops empty-form spam; the 4000-char ceiling keeps
            abuse cheap.
        evidenceUrl:
          type: string
          minLength: 1
          maxLength: 2000
          nullable: true
          description: Optional supporting URL (study, manufacturer page, …).
        email:
          type: string
          maxLength: 320
          nullable: true
          description: |
            Optional reply-to so the editors can attribute the catch and
            send confirmation when the fix ships, per the policy
            ("with attribution to the reader if one prompted it").
        website:
          type: string
          maxLength: 200
          nullable: true
          description: |
            Honeypot field. Real browsers leave this blank because the
            input is visually hidden via CSS; bots that scrape the form
            and fill every field will populate it. Submissions where
            this is non-empty are silently accepted but never persisted.
      required:
        - pageUrl
        - description
    CorrectionRecord:
      type: object
      description: |
        A reader-submitted correction as the editor inbox sees it.
        Includes the same fields the reader posted plus the workflow
        state (`status`, `internalNote`) and timestamps the inbox
        renders the SLA clock from.
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
          description: When the reader hit Submit. The SLA clock starts here.
        updatedAt:
          type: string
          format: date-time
          description: |
            Last time an editor PATCHed the row. Equal to `createdAt`
            for untouched submissions.
        pageUrl:
          type: string
        description:
          type: string
        evidenceUrl:
          type: string
          nullable: true
        submitterEmail:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, triaged, applied, dismissed]
        internalNote:
          type: string
          nullable: true
          description: |
            Editor-only scratchpad. Never surfaced to readers; only
            visible inside the inbox UI.
      required:
        - id
        - createdAt
        - updatedAt
        - pageUrl
        - description
        - evidenceUrl
        - submitterEmail
        - userAgent
        - status
        - internalNote
    EditorCorrectionsList:
      type: object
      properties:
        status:
          type: string
          enum: [new, triaged, applied, dismissed, all]
          description: The status filter the response is scoped to.
        corrections:
          type: array
          items:
            $ref: "#/components/schemas/CorrectionRecord"
      required:
        - status
        - corrections
    UpdateCorrectionBody:
      type: object
      description: |
        At least one of `status` / `internalNote` must be supplied.
        Send `internalNote: null` to clear an existing note.
      properties:
        status:
          type: string
          enum: [new, triaged, applied, dismissed]
          description: New workflow state for the row.
        internalNote:
          type: string
          maxLength: 4000
          nullable: true
          description: |
            Replacement value for the editor-only note. Set to `null`
            to clear an existing note.
...
    ShelfPageKind:
      type: string
      enum: [ingredient, routine, concern]
      description: |
        Which kind of editorial page surfaced the shelf card. Combined
        with `pageSlug`, this identifies the source page so analytics
        can be sliced by page kind, by individual page, or both.
    ShelfClickEvent:
      type: object
      properties:
        pageKind:
          $ref: "#/components/schemas/ShelfPageKind"
        pageSlug:
          type: string
          minLength: 1
          maxLength: 200
          description: |
            Per-kind identifier for the source page. Ingredient briefs
            use the catalogue ingredient slug (e.g. `tretinoin`); routine
            and concern pages use the detail-page component file name
            (e.g. `RoutineBareMinimum`, `ConcernRosacea`) because
            multiple catalogue slugs can alias the same component.
        brand:
          type: string
          minLength: 1
          maxLength: 200
        productName:
          type: string
          minLength: 1
          maxLength: 400
        href:
          type: string
          minLength: 1
          maxLength: 2000
      required:
        - pageKind
        - pageSlug
        - brand
        - productName
        - href
    ShelfClickBatch:
      type: object
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/ShelfClickEvent"
      required:
        - events
...